requirements.txt
aiohttp==3.7.3
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
colorama==0.4.4
discord.py==1.6.0
idna==2.10
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==5.1.0
praw==7.1.0
prawcore==1.5.0
pylint==2.6.0
python-dotenv==0.15.0
requests==2.25.1
six==1.15.0
TextToOwO==2.0.1
toml==0.10.2
typing-extensions==3.7.4.3
update-checker==0.18.0
urllib3==1.26.2
websocket-client==0.57.0
wrapt==1.12.1
yarl==1.6.3
